+++ examples/xml/port_interface/client_server_interface.py
@@ -4,6 +4,7 @@ Sender-receiver port interface examples
 import os
 import autosar
 import autosar.xml.element as ar_element
+import autosar.xml.enumeration as ar_enum
 
 
 def create_platform_types(packages: dict[str, ar_element.Package]):
@@ -26,36 +27,17 @@ def create_platform_types(packages: dict[str, ar_element.Package]):
     packages["PlatformImplementationDataTypes"].append(uint32_impl_type)
 
 
-def create_implementation_data_types(packages: dict[str, ar_element.Package]):
-    """
-    Creates non-platform implementation data types
-    """
-    sw_data_def_props = ar_element.SwDataDefPropsConditional(base_type_ref="AUTOSAR_Platform/BaseTypes/uint8")
-    inactive_active_t = ar_element.ImplementationDataType("InactiveActive_T",
-                                                          category="VALUE",
-                                                          sw_data_def_props=sw_data_def_props)
-    packages["ImplementationDataTypes"].append(inactive_active_t)
-
-
-def create_sender_receiver_interface_with_one_element(packages: dict[str, ar_element.Package]):
+def create_port_interfaces(packages: dict[str, ar_element.Package]):
     """
     Creates interface with one element
     """
-    inactive_active_t = packages["ImplementationDataTypes"].find("InactiveActive_T")
-    portinterface = ar_element.SenderReceiverInterface("HeadLightStatus_I")
-    portinterface.make_data_element("HeadLightStatus", type_ref=inactive_active_t.ref())
-    packages["PortInterfaces"].append(portinterface)
-
-
-def create_sender_receiver_interface_with_two_elements(packages: dict[str, ar_element.Package]):
-    """
-    Creates interface with two elements
-    """
-    inactive_active_t = packages["ImplementationDataTypes"].find("InactiveActive_T")
-    portinterface = ar_element.SenderReceiverInterface("InterfaceName")
-    portinterface.make_data_element("Element1", type_ref=inactive_active_t.ref())
-    portinterface.make_data_element("Element2", type_ref=inactive_active_t.ref())
-    packages["PortInterfaces"].append(portinterface)
+    uint32_impl_type = packages["PlatformImplementationDataTypes"].find("uint32")
+    interface = ar_element.ClientServerInterface("FreeRunningTimer_I", is_service=True)
+    operation = interface.make_operation("GetTimeStamp")
+    operation.make_out_argument("value",
+                                ar_enum.ServerArgImplPolicy.USE_ARGUMENT_TYPE,
+                                type_ref=uint32_impl_type.ref())
+    packages["PortInterfaces"].append(interface)
 
 
 def save_xml_files(workspace: autosar.xml.Workspace):
@@ -63,13 +45,10 @@ def save_xml_files(workspace: autosar.xml.Workspace):
     Saves workspace as XML documents
     """
     interface_document_path = os.path.abspath(os.path.join(os.path.dirname(
-        __file__), 'data', 'sender_receiver_interface.arxml'))
-    datatype_document_path = os.path.abspath(os.path.join(os.path.dirname(
-        __file__), 'data', 'datatypes.arxml'))
+        __file__), 'data', 'client_server_interface.arxml'))
     platform_document_path = os.path.abspath(os.path.join(os.path.dirname(
         __file__), 'data', 'platform.arxml'))
     workspace.create_document(interface_document_path, packages="/PortInterfaces")
-    workspace.create_document(datatype_document_path, packages="/DataTypes")
     workspace.create_document(platform_document_path, packages="/AUTOSAR_Platform")
     workspace.write_documents()
 
@@ -81,16 +60,12 @@ def main():
     workspace = autosar.xml.Workspace()
     packages = dict(zip(["PlatformBaseTypes",
                          "PlatformImplementationDataTypes",
-                         "ImplementationDataTypes",
                          "PortInterfaces"],
                     workspace.make_packages("AUTOSAR_Platform/BaseTypes",
                                             "AUTOSAR_Platform/ImplementationDataTypes",
-                                            "DataTypes/ImplementationDataTypes",
                                             "PortInterfaces")))
     create_platform_types(packages)
-    create_implementation_data_types(packages)
-    create_sender_receiver_interface_with_one_element(packages)
-    create_sender_receiver_interface_with_two_elements(packages)
+    create_port_interfaces(packages)
     save_xml_files(workspace)